Fix test case imports and refactor
