+++ microservices/scraper_australia/app/tests/scraper_australia_test.py
@@ -1,0 +1,11 @@
+import pytest
+import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join('..', 'microservices', 'scraper_australia')))
+
+from app import scraper
+
+
+def test_basic():
+    assert scraper.get_page_data() is not None
