+++ microservices/scraper_example/app/tests/scraper_example_test.py
@@ -1,0 +1,11 @@
+import pytest
+import sys
+import os
+
+sys.path.append(os.path.abspath(os.path.join('..', 'microservices')))
+
+from scraper_example.app import scraper
+
+
+def test_basic():
+    assert scraper.get_page_data() is not None
