--- microservices/tests/scraper_australia_test.py
@@ -1,6 +1,0 @@
-import pytest
-from microservices.scraper_australia.app import scraper
-
-
-def test_basic():
-    assert scraper.get_page_data() is not None
