--- microservices/tests/scraper_hati_malaysia_test.py
@@ -1,6 +1,0 @@
-import pytest
-from microservices.scraper_hati_malaysia.app import scraper
-
-
-def test_basic():
-    assert scraper.scrape(one=True) is not None
