--- microservices/tests/scraper_ngolist_test.py
@@ -1,6 +1,0 @@
-import pytest
-from microservices.scraper_ngolist.app import scraper
-
-
-def test_basic():
-    assert scraper.basepage_scrape() is not None
