--- microservices/tests/scraper_ngothai_test.py
@@ -1,6 +1,0 @@
-import pytest
-from microservices.scraper_ngothai.app import scraper
-
-
-def test_basic():
-    assert scraper.get_one_ngo() is not None
