--- microservices/tests/scraper_ngouganda_test.py
@@ -1,6 +1,0 @@
-import pytest
-from microservices.scraper_ngouganda.app import scraper
-
-
-def test_basic():
-    assert scraper.get_one() is not None
